encoding image data...
It seems that tf wants the image stored in this file, rather than the way that the image is stored in this file...

We're already reading the image into img. I think we can just leave an encoded version of that file in the image/encoded variable. 

i.e. 
encoded_image_data = tf.image. encode_png(img)